+++ src/components/vessels/vessels-list/vessels-list.vue
@@ -1,0 +1,3 @@
+<template>
+  <h2>Vessels list</h2>
+</template>
