+++ src/components/welcome/welcome.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-  import PositiveButton from '../../components/atoms/buttons/positive.vue'
+  import PositiveButton from '../atoms/buttons/positive.vue'
 
   export default {
     components: {
