+++ src/pages/auth/signin.vue
@@ -45,7 +45,7 @@
 <script>
   import { required, email } from 'vuelidate/lib/validators'
 
-  import FormWrapper from '../form-wrapper.vue'
+  import FormWrapper from '../../components/form-wrapper.vue'
   import PositiveButton from '../../components/atoms/buttons/positive.vue'
 
   export default {
@@ -89,8 +89,8 @@
     },
 
     components: {
-      'positive-button': PositiveButton,
-      'form-wrapper': FormWrapper
+      PositiveButton,
+      FormWrapper
     }
   }
 </script>
