+++ src/pages/vessels.vue
@@ -1,10 +1,18 @@
 <template>
   <div id="vessels">
     <h1>Vessels</h1>
+    <vessels-list></vessels-list>
   </div>
 </template>
 
 <script>
+  import vesselsList from '../components/vessels/vessels-list/vessels-list.vue'
+
+  export default {
+    components: {
+      vesselsList
+    }
+  }
 </script>
 
 <style scoped lang="scss">
